control max hamm to cap perc
keep the max hamming value in avg calcs to the top 'matching'
value, like 5 for PHash. This keeps the algorithm a bit more
objective and solid so that the larger values that all become
'it's not matching' vs. any degree of difference are accounted
for better. The lower values seemed to skew from higher hamming
frames which anything 5 or greater should be an equal quality hit.
this allows the 20 * 5 to max out at 100 for PHash. other algorithms
are accounted for, although still just research exploring the
behavior.